+++ Source/PropertyTools.Wpf.Tests/ItemsBag/ItemsBagTests.cs
@@ -392,6 +392,213 @@ namespace PropertyTools.Wpf.Tests
             Assert.That(p1.ShouldSerializeValue(bag), Is.False);
         }
 
+        [Test]
+        public void PropertyType_ForIntValueType_ReturnsNullableInt()
+        {
+            var t0 = new TestObject();
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("IntValue", false);
+
+            Assert.That(p1.PropertyType, Is.EqualTo(typeof(int?)));
+        }
+
+        [Test]
+        public void PropertyType_ForDoubleValueType_ReturnsNullableDouble()
+        {
+            var t0 = new TestObject();
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("DoubleValue", false);
+
+            Assert.That(p1.PropertyType, Is.EqualTo(typeof(double?)));
+        }
+
+        [Test]
+        public void PropertyType_ForEnumValueType_ReturnsNullableEnum()
+        {
+            var t0 = new TestObject();
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("EnumValue", false);
+
+            Assert.That(p1.PropertyType, Is.EqualTo(typeof(TestEnum?)));
+        }
+
+        [Test]
+        public void GetValue_IntValueWithDifferentValues_ReturnsNull()
+        {
+            var t0 = new TestObject() { IntValue = 10 };
+            var t1 = new TestObject() { IntValue = 20 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("IntValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(null));
+        }
+
+        [Test]
+        public void GetValue_IntValueWithEqualValues_ReturnsValue()
+        {
+            var t0 = new TestObject() { IntValue = 42 };
+            var t1 = new TestObject() { IntValue = 42 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("IntValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(42));
+        }
+
+        [Test]
+        public void GetValue_DoubleValueWithDifferentValues_ReturnsNull()
+        {
+            var t0 = new TestObject() { DoubleValue = 1.5 };
+            var t1 = new TestObject() { DoubleValue = 2.5 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("DoubleValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(null));
+        }
+
+        [Test]
+        public void GetValue_DoubleValueWithEqualValues_ReturnsValue()
+        {
+            var t0 = new TestObject() { DoubleValue = 3.14 };
+            var t1 = new TestObject() { DoubleValue = 3.14 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("DoubleValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(3.14));
+        }
+
+        [Test]
+        public void GetValue_EnumValueWithDifferentValues_ReturnsNull()
+        {
+            var t0 = new TestObject() { EnumValue = TestEnum.First };
+            var t1 = new TestObject() { EnumValue = TestEnum.Second };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("EnumValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(null));
+        }
+
+        [Test]
+        public void GetValue_EnumValueWithEqualValues_ReturnsValue()
+        {
+            var t0 = new TestObject() { EnumValue = TestEnum.Third };
+            var t1 = new TestObject() { EnumValue = TestEnum.Third };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("EnumValue", false);
+
+            Assert.That(p1.GetValue(bag), Is.EqualTo(TestEnum.Third));
+        }
+
+        [Test]
+        public void SetValue_IntValueWithMultipleObjects_SetsAllValues()
+        {
+            var t0 = new TestObject() { IntValue = 10 };
+            var t1 = new TestObject() { IntValue = 20 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("IntValue", false);
+
+            p1.SetValue(bag, 100);
+
+            Assert.That(t0.IntValue, Is.EqualTo(100));
+            Assert.That(t1.IntValue, Is.EqualTo(100));
+        }
+
+        [Test]
+        public void SetValue_DoubleValueWithMultipleObjects_SetsAllValues()
+        {
+            var t0 = new TestObject() { DoubleValue = 1.5 };
+            var t1 = new TestObject() { DoubleValue = 2.5 };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("DoubleValue", false);
+
+            p1.SetValue(bag, 9.99);
+
+            Assert.That(t0.DoubleValue, Is.EqualTo(9.99));
+            Assert.That(t1.DoubleValue, Is.EqualTo(9.99));
+        }
+
+        [Test]
+        public void SetValue_EnumValueWithMultipleObjects_SetsAllValues()
+        {
+            var t0 = new TestObject() { EnumValue = TestEnum.First };
+            var t1 = new TestObject() { EnumValue = TestEnum.Second };
+            var bag = new ItemsBag(new[] { t0, t1 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("EnumValue", false);
+
+            p1.SetValue(bag, TestEnum.Third);
+
+            Assert.That(t0.EnumValue, Is.EqualTo(TestEnum.Third));
+            Assert.That(t1.EnumValue, Is.EqualTo(TestEnum.Third));
+        }
+
+        [Test]
+        public void SetValue_NullOnNonNullableIntValue_ThrowsTargetException()
+        {
+            var t0 = new TestObject() { IntValue = 10 };
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("IntValue", false);
+
+            // Setting null on a non-nullable value type should throw
+            Assert.Throws<System.Reflection.TargetException>(() => p1.SetValue(bag, null));
+            // Original value should remain unchanged
+            Assert.That(t0.IntValue, Is.EqualTo(10));
+        }
+
+        [Test]
+        public void SetValue_NullOnNonNullableDoubleValue_ThrowsTargetException()
+        {
+            var t0 = new TestObject() { DoubleValue = 1.5 };
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("DoubleValue", false);
+
+            // Setting null on a non-nullable value type should throw
+            Assert.Throws<System.Reflection.TargetException>(() => p1.SetValue(bag, null));
+            // Original value should remain unchanged
+            Assert.That(t0.DoubleValue, Is.EqualTo(1.5));
+        }
+
+        [Test]
+        public void SetValue_NullOnNonNullableEnumValue_ThrowsTargetException()
+        {
+            var t0 = new TestObject() { EnumValue = TestEnum.First };
+            var bag = new ItemsBag(new[] { t0 });
+            var provider = new ItemsBagTypeDescriptionProvider();
+            var td = provider.GetTypeDescriptor(typeof(ItemsBag), bag);
+            var p1 = td.GetProperties().Find("EnumValue", false);
+
+            // Setting null on a non-nullable value type should throw
+            Assert.Throws<System.Reflection.TargetException>(() => p1.SetValue(bag, null));
+            // Original value should remain unchanged
+            Assert.That(t0.EnumValue, Is.EqualTo(TestEnum.First));
+        }
+
         private class TestObject
         {
             public bool IsChecked => this.Checked;
@@ -399,6 +606,17 @@ namespace PropertyTools.Wpf.Tests
             public string Name { get; set; }
             public int? NullableInt { get; set; }
             public GenericStruct<int> GenericValue { get; set; }
+            public int IntValue { get; set; }
+            public double DoubleValue { get; set; }
+            public TestEnum EnumValue { get; set; }
+        }
+
+        private enum TestEnum
+        {
+            None,
+            First,
+            Second,
+            Third
         }
 
         private class ObservableTestObject : System.ComponentModel.INotifyPropertyChanged